OkHttp: Merge all session services into one
