--- sessions-core/src/main/java/com/adyen/checkout/sessions/api/SessionDetailsService.kt
@@ -1,49 +1,0 @@
-/*
- * Copyright (c) 2022 Adyen N.V.
- *
- * This file is open source and available under the MIT license. See the LICENSE file for more info.
- *
- * Created by josephj on 18/3/2022.
- */
-
-package com.adyen.checkout.sessions.api
-
-import com.adyen.checkout.core.api.Environment
-import com.adyen.checkout.core.api.HttpClientFactory
-import com.adyen.checkout.core.log.LogUtil
-import com.adyen.checkout.core.log.Logger
-import com.adyen.checkout.core.model.toStringPretty
-import com.adyen.checkout.sessions.model.payments.SessionDetailsRequest
-import com.adyen.checkout.sessions.model.payments.SessionDetailsResponse
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.withContext
-import org.json.JSONObject
-
-private val TAG = LogUtil.getTag()
-
-internal class SessionDetailsService(
-    private val environment: Environment
-) {
-
-    suspend fun submitDetails(
-        request: SessionDetailsRequest,
-        sessionId: String,
-        clientKey: String,
-    ): SessionDetailsResponse = withContext(Dispatchers.IO) {
-        val path = "v1/sessions/$sessionId/paymentDetails?clientKey=$clientKey"
-
-        Logger.v(TAG, "call - $path")
-
-        val requestJson = SessionDetailsRequest.SERIALIZER.serialize(request)
-
-        Logger.v(TAG, "request - ${requestJson.toStringPretty()}")
-
-        val httpClient = HttpClientFactory.getHttpClient(environment.baseUrl)
-        val result = httpClient.post(path, requestJson.toString())
-        val resultJson = JSONObject(String(result, Charsets.UTF_8))
-
-        Logger.v(TAG, "response: ${resultJson.toStringPretty()}")
-
-        SessionDetailsResponse.SERIALIZER.deserialize(resultJson)
-    }
-}
